Module Libsecp256k1.Internal

module Num : sig ... end

module Scalar : sig ... end

module Field : sig ... end

Field element module. * * Field elements can be represented in several ways, but code accessing * it (and implementations) need to take certain properties into account: * - Each field element can be normalized or not. * - Each field element has a magnitude, which represents how far away * its representation is away from normalization. Normalized elements * always have a magnitude of 1, but a magnitude of 1 doesn't imply * normality.

module Group : sig ... end